hours filter works
