+++ lib/filterLocations.js
@@ -1,9 +1,30 @@
 import R from 'ramda'
 import { relevantTaxonomies } from './taxonomies'
 
-const date = new Date()
-const milTime = date.getHours() * 100 + date.getMinutes()
 const weekdays = ['Sunday','Monday','Tuesday','Wednesday','Thursday','Friday','Saturday']
+
+
+export const isOpen = (services) => {
+  const now = new Date()
+  const currentTime = now.getHours() * 100 + now.getMinutes()
+  const currentDay = weekdays[now.getDay()].toLowerCase()
+
+  const isBetween = (num, min, max) => (
+    num > min && num < max
+  )
+
+  const scheduleContainsCurrentTime = (schedule) => (
+    currentDay === schedule.weekday.toLowerCase() &&
+      isBetween(currentTime, schedule.opensAt, schedule.closesAt)
+  )
+
+  const serviceHasMatchingSchedule = (service) => {
+    if (!service.schedules) return true
+    return (service.schedules).some(scheduleContainsCurrentTime)
+  }
+
+  return Object.values(services).some(serviceHasMatchingSchedule)
+}
 
 // Types!
 //
@@ -82,15 +103,14 @@ function genderFilter(locationOptions) {
 // LocationFilterOptions -> LocationFilterOptions
 function hoursFilter(locationOptions) {
   const { location, options, isValid } = locationOptions
+  console.log(location.name)
   if (!isValid || options.hours === 'all') return locationOptions
   const { services } = location
 
-  return Object.values(services).some(service =>
-  	 service.schedules && service.schedules.some(schedule =>
-     	 schedule.opens_at <= milTime &&
-      			 schedule.closes_at >= milTime &&
-             schedule.weekday === weekdays[date.getDay()])
-  )
+  const isStillValid = isOpen(services)
+  console.log(isStillValid)
+
+  return {...locationOptions, isValid: isStillValid }
 }
 
 // LocationFilterOptions -> Bool
